+++ cypress/integration/Input/index.spec.js
@@ -55,6 +55,11 @@ describe('Input', () => {
         .type('test text')
         .should('have.value', 'test ');
     });
+    it('MaxLength should limit the number of characters', () => {
+      cy.get('input#only5Characters')
+        .type('test text')
+        .should('have.value', 'test ');
+    });
     it('Should display characters equivalent to those entered', () => {
       cy.get('input#corr-Input')
         .type('888888899879465143164651356')
@@ -142,6 +147,21 @@ describe('Input', () => {
         .get('.input-clear-icon')
         .click()
       cy.focused().should('have.attr', 'id', 'hasClearButton')
+    });
+    describe('Paste event + maxLength attribute', () => {
+      it('Should paste value adjusted to maxLength, if input is empty', () => {
+        cy.get('input#only5Characters')
+          .paste('test text')
+          .should('have.value', 'test ');
+      });
+      it('Should paste value adjusted to maxLength, if old value is replaced completely', () => {
+        cy.get('input#only5Characters')
+          .paste('test text')
+          .should('have.value', 'test ')
+          .type('{selectall}')
+          .paste('888888888888')
+          .should('have.value', '88888')
+      });
     });
   });
 });